각 패스워드 요소에 자동완성 방지 추가

+++ src/component/auth/login.js
@@ -1,20 +1,20 @@
-import "../../css/auth.css";
-import { useEffect, useState } from "react";
-import { useNavigate } from "react-router-dom";
-import axios from "axios";
-import jwtDecode from "jwt-decode";
+import '../../css/auth.css';
+import { useEffect, useState } from 'react';
+import { useNavigate } from 'react-router-dom';
+import axios from 'axios';
+import jwtDecode from 'jwt-decode';
 
 const Login = () => {
   const navigate = useNavigate();
   const goRegister = () => {
-    navigate("/register");
+    navigate('/register');
   };
   const goHome = () => {
-    navigate("/");
+    navigate('/');
   };
 
-  const [id, setId] = useState("");
-  const [password, setPassword] = useState("");
+  const [id, setId] = useState('');
+  const [password, setPassword] = useState('');
 
   // useEffect(() => {
   //   axios
@@ -36,49 +36,56 @@ const Login = () => {
     setPassword(e.target.value);
   };
 
+  useEffect(() => {
+    if (localStorage.getItem('token') && window.confirm('잘못된 접근입니다.')) {
+      window.location.replace('/');
+    }
+  }, []);
+
   const login = () => {
     const userLoginInfo = {};
     userLoginInfo.id = id;
     userLoginInfo.password = password;
     axios
-      .post("http://localhost:3001/api/authApiData/login", userLoginInfo)
+      .post('http://localhost:3001/api/authApiData/login', userLoginInfo)
       .then((res) => {
         console.log(res);
         //로그인 성공 했을 때
         if (res.data.auth === true) {
-          localStorage.setItem("token", res.data.refreshToken);
+          localStorage.setItem('token', res.data.refreshToken);
           console.log(res.data);
-          const token = localStorage.getItem("token");
+          const token = localStorage.getItem('token');
           console.log(token);
           const RefreshPayload = jwtDecode(token);
           const accessPayload = jwtDecode(RefreshPayload.accessToken);
 
           axios.defaults.headers.common[
-            "Authorization"
+            'Authorization'
           ] = `${res.data.refreshToken}`; //앞으로 api통신에 토큰이 들어가있음
           goHome();
         }
         // 비밀번호가 틀렸을 때
         else if (res.data.authPassword === false) {
-          alert("비밀번호가 틀렸습니다.");
+          alert('비밀번호가 틀렸습니다.');
         }
         // 아이디가 틀렸을 때
         else if (res.data.auth === false) {
-          alert("아이디가 틀렸습니다.");
+          alert('아이디가 틀렸습니다.');
         }
       });
   };
 
   return (
     <>
-      <div className="login_container auth">
-        <div className="login_box">
-          <div className="login_input">
+      <div className='login_container auth'>
+        <div className='login_box'>
+          <div className='login_input'>
             <p>아이디</p>
-            <input name="id" value={id} onChange={onChangeId}></input>
+            <input name='id' value={id} onChange={onChangeId}></input>
             <p>비밀번호</p>
             <input
-              name="password"
+              autocomplete='off'
+              name='password'
               value={password}
               onChange={onChangePassword}
             ></input>
@@ -87,7 +94,7 @@ const Login = () => {
             <button onClick={login}>로그인</button>
             <p>
               계정이 없으신가요?
-              <a href="#" onClick={goRegister}>
+              <a href='#' onClick={goRegister}>
                 회원가입
               </a>
             </p>
